backend/pyproject.toml
[project]
name = "the-website-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.29,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-cov (>=5.0.0,<6.0.0)",
    "pytest-asyncio (>=0.24.0,<1.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy-serializer (>=1.4.22,<2.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "stripe (>=12.1.0,<13.0.0)",
    "boto3 (>=1.38.13,<2.0.0)",
    "google-api-python-client (>=2.172.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "groq (>=0.31.0,<0.32.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "huggingface-hub (>=0.34.4,<0.35.0)",
    "mistralai (>=1.9.3,<2.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
]

[tool.poetry]
packages = [
    { include = "api", from = "." },
    { include = "tests", from = "." }
]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
